+++ src/__tests__/App.test.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 
-import { App } from './App';
+import { App } from '../App';
 
 describe('App Tests', () => {
   it('renders without crashing', () => {
