--- test/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  rootDir: '../',
-  preset: 'ts-jest',
-};
